Refactor/upsert

+++ src/server/api/routers/private/institutions.router.ts
@@ -1,44 +1,27 @@
 import { createTRPCRouter, protectedProcedure } from "@/server/api/trpc";
 import { TRPCError } from "@trpc/server";
-import {
-  institutionSchema,
-  idSchema,
-  updateInstitutionSchema,
-} from "@/server/api/schemas";
+import { idSchema, updateInstitutionSchema } from "@/server/api/schemas";
+import { PermissionTypes, findAndValidatePermission } from "../../auth";
 
 export const institutionRouter = createTRPCRouter({
-  createOne: protectedProcedure
-    .input(institutionSchema)
-    .mutation(async ({ ctx, input }) => {
-      const existentCompany = await ctx.prisma.institution.findUnique({
-        where: { company_code: input.company_code },
-      });
-
-      if (existentCompany) {
-        throw new TRPCError({
-          code: "CONFLICT",
-          message: `Institution ${existentCompany.name} already exists`,
-        });
-      }
-      return await ctx.prisma.institution.create({
-        data: {
-          company_code: input.company_code,
-          name: input.name,
-          email: input.email,
-          attributes: input.attributes,
-          additional_info: input.additional_info,
-          static_logo_url: input.static_logo_url,
-        },
-      });
-    }),
   findById: protectedProcedure.input(idSchema).query(async ({ ctx, input }) => {
+    const institutionManagementPermission = findAndValidatePermission(
+      PermissionTypes.INSTITUTION_MANAGEMENT,
+      ctx.session.user.permissions
+    );
+    if (
+      !institutionManagementPermission ||
+      !institutionManagementPermission.value
+    )
+      throw new TRPCError({ code: "UNAUTHORIZED" });
     const institution = await ctx.prisma.institution.findUnique({
       where: { id: input.id },
     });
+
     if (!institution) {
       throw new TRPCError({
         code: "NOT_FOUND",
-        message: "Company not found",
+        message: "Institution not found",
       });
     }
     return institution;
@@ -52,17 +35,44 @@ export const institutionRouter = createTRPCRouter({
   delete: protectedProcedure
     .input(idSchema)
     .mutation(async ({ ctx, input }) => {
+      const institutionManagementPermission = findAndValidatePermission(
+        PermissionTypes.INSTITUTION_MANAGEMENT,
+        ctx.session.user.permissions
+      );
+      if (
+        !institutionManagementPermission ||
+        !institutionManagementPermission.value
+      )
+        throw new TRPCError({ code: "UNAUTHORIZED" });
       await ctx.prisma.institution.delete({
         where: { id: input.id },
       });
       return true;
     }),
-  update: protectedProcedure
+  upsert: protectedProcedure
     .input(updateInstitutionSchema)
     .mutation(async ({ ctx, input }) => {
-      const institution = await ctx.prisma.institution.update({
-        where: { id: input.id },
-        data: {
+      const institutionManagementPermission = findAndValidatePermission(
+        PermissionTypes.INSTITUTION_MANAGEMENT,
+        ctx.session.user.permissions
+      );
+      if (
+        !institutionManagementPermission ||
+        !institutionManagementPermission.value
+      )
+        throw new TRPCError({ code: "UNAUTHORIZED" });
+
+      const institution = await ctx.prisma.institution.upsert({
+        where: { company_code: input.company_code, id: input.id },
+        create: {
+          company_code: input.company_code,
+          name: input.name,
+          email: input.email,
+          attributes: input.attributes,
+          additional_info: input.additional_info,
+          static_logo_url: input.static_logo_url,
+        },
+        update: {
           company_code: input.company_code,
           name: input.name,
           additional_info: input.additional_info,
